requirements.txt
asgiref==3.3.4
Django==3.2.4
django-ckeditor==6.1.0
django-js-asset==1.2.2
gunicorn==20.1.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
Django==3.2.4
django-ckeditor==6.1.0
django-js-asset==1.2.2
gunicorn==20.1.0
pdfkit==0.6.1
Pillow==8.2.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-ckeditor==6.1.0
django-js-asset==1.2.2
gunicorn==20.1.0
idna==2.10
pdfkit==0.6.1
Pillow==8.2.0
PyPDF2==1.26.0
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
